Changed calibration file format.
Calibration file now records the number of points in a spectrum and a pair of
photon energies and corresponding pixel indices on the CSPAD. This makes it
human-readable and editable.

Also did some refactoring to make the code more easily modified to accommodate
automatic calibration schemes other than k alpha/k beta.