Sketched out config file system.
Running the 'init' sub-command of datac.py initalizes config.py
in the user's local directory, which the user must then edit.

Also began consolidation of all the scripts in this package into
sub-commands under a top-level script. Still in progress; I need to
move g_spreadsheet_sync.py into a module and add the XES script as
a sub-command.