Logbook data is now downloaded automatically and transfered via ZMQ sockets.
logbook.py gets data from a google spreadsheet and pushes it to
clients using a ZMQ pub socket. This change (from the text file-
based exchange of logbook information) was motivated in part by the
need to associate additional data to run group labels.

Started implementing a function to plot powder peak intensity progressions
(as a function of beam transmission). Need to add in a parameter for the
focal spot size.