New plotting modes in xrd.py; tentative work on automatic batch job submission.
xrd.py now plots the progression of Bragg peak intensities as a
function of incident flux density (based on transmission and on
focal spot size, both of which are specified in the logbook).

Also attempted to write a decorator for automatic generation and
submission of batch jobs, but the batch scripts mysteriously fail
when called from within python using os.system.

Reorganized handling of detector parameters in config.py.