Implemented faster generation of text and plots for pre-computed commands.
This is done using the new module database.py, which provides a decorator
for selectively storing and re-executing individual functions without rerunning
the code that precedes them.

A quirk of using it is that the decorated functions cannot be passed any
np.ndarrays as arguments, due to what appears to be a bug/limitation of
dill.